requirements.txt
# LangGraph / LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0

# Vector DB
chromadb>=0.5.0

# Embedding
sentence-transformers>=3.0.0

# LLM (Ollama)
ollama>=0.3.0

# Crawling
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Data Processing
pyyaml>=6.0.0

# API Server
fastapi>=0.115.0
uvicorn>=0.32.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
